lib: rename curlx/fopen.[ch] to curlx/win32-fopen.[ch]
To make it more apparent what the file is for and separate it better
from curl_fopen.[ch]
