+++ lib/curlx/win32-fopen.h
