+++ lib/file.c
@@ -59,7 +59,7 @@
 #include "transfer.h"
 #include "url.h"
 #include "parsedate.h" /* for the week day and month names */
-#include "curlx/fopen.h"
+#include "curlx/win32-fopen.h"
 #include "curl_range.h"
 
 #if defined(_WIN32) || defined(MSDOS)
