+++ lib/memdebug.c
@@ -29,7 +29,7 @@
 
 #include "urldata.h"
 #include "curl_threads.h"
-#include "curlx/fopen.h"  /* for CURLX_FOPEN_LOW(), CURLX_FREOPEN_LOW() */
+#include "curlx/win32-fopen.h"  /* for CURLX_FOPEN_*() */
 
 #ifdef USE_BACKTRACE
 #include <backtrace.h>
