+++ lib/vtls/vtls.c
@@ -60,7 +60,7 @@
 #include "strcase.h"
 #include "url.h"
 #include "progress.h"
-#include "curlx/fopen.h"
+#include "curlx/win32-fopen.h"
 #include "curl_sha256.h"
 #include "curlx/base64.h"
 #include "curlx/inet_pton.h"
