+++ tests/perf/first.h
@@ -33,7 +33,7 @@
 #include "curl_setup.h"
 
 #include "curlx/base64.h" /* for curlx_base64* */
-#include "curlx/fopen.h" /* for curlx_f*() */
+#include "curlx/win32-fopen.h" /* for curlx_f*() */
 #include "curlx/strparse.h" /* for curlx_str_* parsing functions */
 #include "curlx/timeval.h" /* for curlx_now type and related functions */
